+++ python_dev_tools/formatters/format_file.py
@@ -1,8 +1,12 @@
 """Definition of format function, calling all formatters."""
 
+from .autoflake import AutoflakeFormatter
 from .black import BlackFormatter
 
-formatters = [BlackFormatter]
+formatters = [
+    AutoflakeFormatter,
+    BlackFormatter,
+]  # BlackFormatter should be the last one
 
 
 def format_file(filepath):